Add ability to give scatterer constraints to the fitter. We handle these constraints for now by giving the minimizer an array of infinities, causing it to avoid that scatterer configuration.
Fixes to get tests passing on new numpy (some warnings about implicit casts)

Fix a test I accidentally broke when adding error for defining a model
with no free parameters

+++ holopy/scattering/theory/mie_f/miescatlib.py
@@ -31,9 +31,9 @@ from numpy import sin, cos, array
 
 def scatcoeffs(m, x, nstop): # see B/H eqn 4.88
     # implement criterion used by BHMIE plus a couple more orders to
-    # be safe 
-    nmx = array([nstop, np.round_(np.absolute(m*x))]).max() + 20
-    Dnmx = mie_specfuncs.log_der_1(m*x, nmx, nstop) 
+    # be safe
+    nmx = int(array([nstop, np.round_(np.absolute(m*x))]).max()) + 20
+    Dnmx = mie_specfuncs.log_der_1(m*x, nmx, nstop)
     n = np.arange(nstop+1)
     psi, xi = mie_specfuncs.riccati_psi_xi(x, nstop)
     psishift = np.concatenate((np.zeros(1), psi))[0:nstop+1]
@@ -44,7 +44,7 @@ def scatcoeffs(m, x, nstop): # see B/H eqn 4.88
 
 def internal_coeffs(m, x, n_max):
     '''
-    Calculate internal Mie coefficients c_n and d_n given 
+    Calculate internal Mie coefficients c_n and d_n given
     relative index, size parameter, and maximum order of expansion.
 
     Follow Bohren & Huffman's convention. Note that van de Hulst and Kerker
@@ -58,29 +58,29 @@ def internal_coeffs(m, x, n_max):
     dl = m * ratio * (D3x - D1x) / (m * D3x - D1mx)
     return array([cl[1:], dl[1:]]) # start from l = 1
 
-def nstop(x): 
+def nstop(x):
     #takes size parameter, outputs order to compute to according to
     # Wiscombe, Applied Optics 19, 1505 (1980).
     # 7/7/08: generalize to apply same criterion when x is complex
-    return np.round_(np.absolute(x+4.05*x**(1./3.)+2))
+    return int(np.round_(np.absolute(x+4.05*x**(1./3.)+2)))
 
 def asymmetry_parameter(al, bl):
     '''
     Inputs: an, bn coefficient arrays from Mie solution
-    
+
     See discussion on Bohren & Huffman p. 120.
     The output of this function omits the prefactor of 4/(x^2 Q_sca).
     '''
     lmax = al.shape[0]
     l = np.arange(lmax) + 1
-    selfterm = (l[:-1] * (l[:-1] + 2.) / (l[:-1] + 1.) * 
-                np.real(al[:-1] * np.conj(al[1:]) + 
+    selfterm = (l[:-1] * (l[:-1] + 2.) / (l[:-1] + 1.) *
+                np.real(al[:-1] * np.conj(al[1:]) +
                         bl[:-1] * np.conj(bl[1:]))).sum()
-    crossterm = ((2. * l + 1.)/(l * (l + 1)) * 
+    crossterm = ((2. * l + 1.)/(l * (l + 1)) *
                  np.real(al * np.conj(bl))).sum()
     return selfterm + crossterm
 
-def cross_sections(al, bl): 
+def cross_sections(al, bl):
     '''
     Calculates scattering and extinction cross sections
     given arrays of Mie scattering coefficients an and bn.
@@ -92,14 +92,12 @@ def cross_sections(al, bl):
     lmax = al.shape[0]
 
     l = np.arange(lmax) + 1
-    prefactor = (2. * l + 1.) 
+    prefactor = (2. * l + 1.)
     cscat = (prefactor * (np.abs(al)**2 + np.abs(bl)**2)).sum()
     cext = (prefactor * np.real(al + bl)).sum()
-    
+
     # see p. 122
     alts = 2. * (np.arange(lmax) % 2) - 1
     cback = np.abs((prefactor * alts * (al - bl)).sum())**2
 
     return array([cscat, cext, cback])
-
-
